ml/requirements.txt
# PyTorch (需要根据 CUDA 版本选择合适的安装命令)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
torch>=2.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.3.0

# 进度条
tqdm>=4.65.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 训练监控
tensorboard>=2.11.0

# ONNX 导出和推理
onnx>=1.14.0
onnxruntime-gpu>=1.16.0  # 或 onnxruntime (CPU 版本)
